task-alloc-sim/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "task-alloc-sim"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "numpy>=1.26",
  "pandas>=2.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
